Fixed a compilation warning.

+++ libraries/lts/source/simulation.cpp
@@ -75,7 +75,7 @@ void simulation::randomly_select_state()
   if (state.source_state.states.size()>1)
   {
     // Generate a random size_t with random distribution. 
-    double random_value=static_cast<double>(m_distrib(m_gen))/std::numeric_limits<std::size_t>::max(); 
+    double random_value=static_cast<double>(m_distrib(m_gen))/static_cast<double>(std::numeric_limits<std::size_t>::max()); 
     random_value=random_value-data::sort_real::value(state.source_state.probabilities[state_number]);
     while (random_value>0)
     {
